requirements.txt
datasets>=2.18.0
datasketch>=1.6.4
psutil>=5.9.0
pyarrow>=14.0.0
pandas>=2.0.0
tqdm>=4.66.0
pyyaml>=6.0
matplotlib>=3.8.0
seaborn>=0.13.0
jupyter>=1.0.0
huggingface_hub>=0.21.0
